tinyAgent/pyproject.toml
[project]
name = "tinyagent"
version = "0.1.0"
description = "Python agent loop"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Fabian"}]
keywords = ["agent", "llm", "openrouter", "streaming"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.28.0",
]

[project.optional-dependencies]
examples = [
    "python-dotenv>=1.0.0",
]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["tinyagent*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "C90", "I", "N", "UP", "B", "SIM"]
ignore = [
    "UP007",   # Allow Union[] for TypeAlias compatibility
    "SIM117",  # Nested with statements can be more readable for async context managers
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["tinyagent"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true
files = ["."]
exclude = '(^|/)(\.venv|scripts/)'
explicit_package_bases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.lint]
max-file-lines = 500

[tool.archlint]
library_root = "tinyagent"
core_modules = [
    "agent",
    "agent_loop",
    "agent_tool_execution",
    "agent_types",
    "openrouter_provider",
    "proxy",
    "proxy_event_handlers",
]
nih_filenames = ["dotenv.py", "six.py", "compat.py"]
